refactor: consolidate bot lifecycle into bot/core/lifecycle.py
Extract duplicated startup, extension loading, and event handler logic
from main.py and bot/api.py into a single shared bot/core/lifecycle.py
module. Also fixes the double get_channel() call bug in the original
main.py on_app_command_error handler.
